Fixed Issue With Tile Selection
Tile selection would draw a different tile when placing on the map.
There's a bug with finding tiles, it will get you to the right window but it won't select certain tiles because of the algorithm. The tile is selected to be able to draw on the map but it just won't be highlighted red.
Not major but will need to fix in future.